+++ sample/src/main/java/com/polidea/rxandroidble/sample/util/HexString.java
@@ -6,11 +6,25 @@ public class HexString {
 
     public static String bytesToHex(byte[] bytes) {
         char[] hexChars = new char[bytes.length * 2];
+
         for (int j = 0; j < bytes.length; j++) {
             int v = bytes[j] & 0xFF;
             hexChars[j * 2] = hexArray[v >>> 4];
             hexChars[j * 2 + 1] = hexArray[v & 0x0F];
         }
+
         return new String(hexChars);
+    }
+
+    public static byte[] hexToBytes(String hexRepresentation) {
+        int len = hexRepresentation.length();
+        byte[] data = new byte[len / 2];
+
+        for (int i = 0; i < len; i += 2) {
+            data[i / 2] = (byte) ((Character.digit(hexRepresentation.charAt(i), 16) << 4)
+                    + Character.digit(hexRepresentation.charAt(i + 1), 16));
+        }
+
+        return data;
     }
 }